requirements.txt
Flask==3.1.2
flask-cors==6.0.1
mysql-connector-python==9.5.0
google-generativeai==0.3.0
python-dotenv==1.0.0
PyJWT==2.8.1
